Super rough initial implementation of a GPS-like system
The "OSRS Positioning System" (OPS) template-matches the
minmap to a region of the world map to determine where the
player is. If it's given a set of "destination" coordinates,
it can figure out how far away those coordinates are from the
player's current position and move there. This was inspired by
Olly's "RSWalk" for Simba.

Signed-off-by: takelley1 <hxyz@protonmail.com>