Add funcs for simulating human behavior
Added human_behavior_rand() to roll for a chance of performing
a "humanlike" action, such as checking the player's stats.

Also added a check_skills() behavior, which will be one of the
funcs that human_behavior_rand() will roll against.

Also refactored open_side_stone() so it checks against the
unopened state of the desired side stone rather than the opened
state. Added a new side stone Vision object to hopefully make
it a little faster.

Signed-off-by: takelley1 <hxyz@protonmail.com>