Add rough outline for logging out within a range of times
Instead of just rolling occasionally for a logout, which could
occur much earlier than desired, logout_rand_range() will
only roll for a logout if the desired minimum period of time
has elapsed. It will also force a logout if the desired
maximum period of time has elapsed.

Also added functionality to logout_rand() to show in the log
the time of day that the break will be over.

Signed-off-by: takelley1 <hxyz@protonmail.com>