Refactor gem dropping in miner func as a for loop
Originally we just used a bunch of if statements to 
check if a non-ore item was to be dropped, which
took up lots of space. Now we bundle each drop_[gem]
True/False statement with each gem's filepath into
a tuple and iterate over all the tuples.

Also fixed some old logic that wasn't refactored
properly.