Refactor input.py into Mouse and Keyboard classes
The input functions were sharing way too many parameters to not
be in a class.

Fixed variable reference issue with the drop_gem configs
and their references in main.py

Removed using keyboard to hit "enter" at the "existing user"
screen of the login menu and just had it click on the button
instead to make unit testing easier.

Signed-off-by: takelley1 <hxyz@protonmail.com>