Add outline for a launch_client option in orient()
Currently the script breaks if you try running it without
the client running. This might be frustrating for the user,
so I added an initial outline for an option to launch the
client if it's not already running. This will require the user
to specify the executable path in the config file on
Windows.

For Linux I'll include a shell script in the repo to launch
the client. The user will still have to download the
dependencies manually though (wget, 7zip, and
openJDK). Might be able to add a distro-check in the
script and install the packages automatically
depending on the user's package manager.